requirements.txt
alembic==1.4.2
appdirs==1.4.4
arrow==0.15.7
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.2.5
attrs==20.2.0
Authlib==0.15.2
Babel==2.8.0
bcrypt==3.2.0
Brlapi==0.6.6
cached-property==1.5.1
certifi==2018.1.18
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1
cupshelpers==1.0
dateparser==0.7.6
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.8
django-scheduler==0.9.3
dnspython==1.16.0
docker==4.3.1
docker-compose==1.27.3
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.13.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.17.1
httplib2==0.11.3
icalendar==4.0.6
idna==2.10
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwt==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.3
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
moment==0.10
numpy==1.19.4
oauth==1.0.1
olefile==0.45.1
paramiko==2.7.2
pbr==5.5.0
pexpect==4.2.1
Pillow==5.1.0
print==1.3.0
protobuf==3.0.0
psycopg2-binary==2.8.5
py3-validate-email==0.2.9
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.3.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.4
python-dateutil==2.6.0
python-debian==0.1.32
python-dotenv==0.14.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2020.1
pywatchman==1.4.0
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
radon==4.3.2
regex==2020.7.14
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
scour==0.36
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
SQLAlchemy==1.3.22
sqlparse==0.3.1
system-service==0.3
systemd-python==234
testresources==2.0.1
texttable==1.6.3
times==0.7
typed-ast==1.4.1
tzlocal==2.1
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
uWSGI==2.0.19.1
validate-email==1.3
vboxapi==1.0
virtualenv==20.0.27
wadllib==1.3.2
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.1
WTForms==2.3.1
xkit==0.0.0
youtube-dl==2018.3.14
zope.interface==4.3.2
